requirements.txt
altgraph==0.17.2
future==0.18.2
macholib==1.15.2
numpy==1.24.3
Pillow==9.5.0
pip==21.2.4
setuptools==58.0.4
six==1.15.0
wheel==0.37.0
